backend/requirements.txt
psycopg2-binary==2.9.7
dj-database-url==2.2.0
dj-email-url==1.0.6
Django==4.2.22
django-cache-url==3.4.5
django-debug-toolbar==3.2.1
django-extensions==4.1
django-phonenumber-field==8.1.0
djangorestframework==3.16.0
drf-standardized-errors==0.14.1
environs==9.3.2
geographiclib==2.0
pillow==11.2.1
python-dotenv==1.1.0
requests==2.32.3
marshmallow==3.19.0
geopy==2.4.1
phonenumberslite==9.0.6
dj-database-url==2.2.0
pygments
rollbar
gunicorn
